+++ src/1-services/devices.services.ts
@@ -1,6 +1,14 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { DeviceMainType, DeviceUpdateType } from '../5-dtos/devices.types';
 import { DevicesRepository } from '../2-repositories/devices.repository';
+import {
+  ExceptionsNames,
+  ResponseToControllersHelper,
+} from '../6-helpers/response.to.controllers.helper';
 
 @Injectable()
 export class DevicesServices {
@@ -19,14 +27,29 @@ export class DevicesServices {
     return this.deviceRepository.changeDevice(deviceUpdate);
   }
 
-  async deleteDevice(deviceId: string) {
-    return this.deviceRepository.deleteDevice(deviceId);
+  async deleteDevice(deviceId: string, userId: string) {
+    const device = await this.deviceRepository.findDeviceById(deviceId);
+    if (!device) {
+      return new ResponseToControllersHelper(
+        true,
+        ExceptionsNames.NotFound_404,
+      );
+    }
+    if (device.userId.toString() != userId) {
+      return new ResponseToControllersHelper(
+        true,
+        ExceptionsNames.Forbidden_403,
+      );
+    }
+    await this.deviceRepository.deleteDevice(deviceId);
+    return new ResponseToControllersHelper(false);
   }
 
   async deleteAllOtherDevices(
     userId: string,
     deviceId: string,
-  ): Promise<boolean> {
-    return this.deviceRepository.deleteAllOtherDevices(userId, deviceId);
+  ): Promise<ResponseToControllersHelper> {
+    await this.deviceRepository.deleteAllOtherDevices(userId, deviceId);
+    return new ResponseToControllersHelper(false);
   }
 }
